Delete confirmation dialog in Blazor

+++ Blazor Tutorial/EmployeeManagement.Web/Pages/DisplayEmployeeBase.cs
@@ -30,16 +30,24 @@ namespace EmployeeManagement.Web.Pages
         [Inject]
         public NavigationManager NavigationManager { get; set; }
 
-        protected async Task Delete_Click()
+        protected PragimTech.Components.ConfirmBase DeleteConfirmation { get; set; }
+
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeID);
-            await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeID);
-            //NavigationManager.NavigateTo("/", true);
+            DeleteConfirmation.Show();
+        }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeID);
+                await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeID);
+            }
         }
 
         protected async Task CheckBoxChanged(ChangeEventArgs e)
-        {
-            IsSelected = (bool)e.Value;
+        {  
             await OnEmployeeSelection.InvokeAsync(IsSelected);
         }
     }
